fix: wrap useSearchParams in Suspense to fix /verify build error
